Complete sign in with apple
