+++ Project 2 AirNote/Controller/Base/LoginViewController.swift
@@ -7,16 +7,10 @@
 
 import UIKit
 import AuthenticationServices
-import Firebase
-import FirebaseAuth
-import CryptoKit
 
 class LoginViewController: UIViewController {
     
     private var signInWithAppleButton = ASAuthorizationAppleIDButton(type: .signIn, style: .black)
-    
-    // Unhashed nonce.
-    fileprivate var currentNonce: String?
     
     // User Manager
     private var userManager = UserManager()
@@ -27,9 +21,23 @@ class LoginViewController: UIViewController {
         configureSignInWithAppleButton()
     }
     
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+    }
+    
 }
 
 extension LoginViewController: ASAuthorizationControllerDelegate, ASAuthorizationControllerPresentationContextProviding {
+    
+    func presentationAnchor(for controller: ASAuthorizationController) -> ASPresentationAnchor {
+        
+        return view.window!
+        
+    }
+    
+    func authorizationController(controller: ASAuthorizationController, didCompleteWithAuthorization authorization: ASAuthorization) {
+        FirebaseSignInManager.shared.authorizationController(controller: controller, didCompleteWithAuthorization: authorization)
+    }
     
     func configureSignInWithAppleButton() {
         
@@ -47,160 +55,20 @@ extension LoginViewController: ASAuthorizationControllerDelegate, ASAuthorizatio
     
     @objc private func signInWithApple() {
         
-        performSignIn()
-        
-    }
-    
-    func performSignIn() {
-        
-        let nonce = randomNonceString()
-        currentNonce = nonce
-        let appleIDProvider = ASAuthorizationAppleIDProvider()
-        let request = appleIDProvider.createRequest()
-        request.requestedScopes = [.fullName, .email]
-        request.nonce = sha256(nonce)
-        
+        let request = FirebaseSignInManager.shared.performSignIn()
         let authorizationController = ASAuthorizationController(authorizationRequests: [request])
         authorizationController.delegate = self
         authorizationController.presentationContextProvider = self
         authorizationController.performRequests()
         
-    }
-    
-    func authorizationController(controller: ASAuthorizationController, didCompleteWithAuthorization authorization: ASAuthorization) {
-        if let appleIDCredential = authorization.credential as? ASAuthorizationAppleIDCredential {
-            guard let nonce = currentNonce else {
-                fatalError("Invalid state: A login callback was received, but no login request was sent.")
-            }
-            guard let appleIDToken = appleIDCredential.identityToken else {
-                print("Unable to fetch identity token")
-                return
-            }
-            guard let idTokenString = String(data: appleIDToken, encoding: .utf8) else {
-                print("Unable to serialize token string from data: \(appleIDToken.debugDescription)")
-                return
-            }
-            // Initialize a Firebase credential.
-            let credential = OAuthProvider.credential(withProviderID: "apple.com",
-                                                      idToken: idTokenString,
-                                                      rawNonce: nonce)
-            // Sign in with Firebase.
-            Auth.auth().signIn(with: credential) { (authResult, error) in
-                if  let user = authResult?.user {
-                    // PresentationStyle
-                    guard let vc = self.storyboard?.instantiateViewController(withIdentifier: "TabBarViewController") as? TabBarViewController else { return }
-                    vc.modalPresentationStyle = .fullScreen
+        FirebaseSignInManager.shared.loginSucess = {
+            
+            guard let vc = UIStoryboard.main.instantiateInitialViewController() else { return }
                     
-                    // Define user's identity
-                    self.userManager.fetchUsers { result in
-                        switch result {
-                        case .success(let users):
-                            self.existingUsers = users
-                            if self.mappingUserInfo(user.uid) == true {
-                                
-                                return
-                                
-                            } else {
-                                
-                                var newUser = user
-                                
-                                
-                                
-                                self.userManager.createUser(user: &, completion: <#T##(Result<String, Error>) -> Void#>)
-                                
-                            }
-                            self.present(vc, animated: true)
-                            
-                        case .failure:
-                            print("Fail to fetch users")
-                        }
-                    }
-                    return
-                }
-            }
-        }
-    }
-    
-    func mappingUserInfo(_ uid: String) -> Bool {
-        
-        var existingUser: User?
-        
-        for user in existingUsers {
-            
-            existingUser = user
+            self.present(vc, animated: true)
             
         }
-        
-        if existingUser != nil {
-            
-            return true
-            
-        } else {
-            
-            return false
-            
-        }
-        
-    }
     
-    func authorizationController(controller: ASAuthorizationController, didCompleteWithError error: Error) {
-        // Handle error.
-        print("Sign in with Apple errored: \(error)")
-    }
-    
-    func presentationAnchor(for controller: ASAuthorizationController) -> ASPresentationAnchor {
-        return view.window!
-    }
-    
-}
-
-// Nonce & sha256 generating function
-extension LoginViewController {
-    
-    // Adapted from https://auth0.com/docs/api-auth/tutorials/nonce#generate-a-cryptographically-random-nonce
-    private func randomNonceString(length: Int = 32) -> String {
-        precondition(length > 0)
-        let charset: [Character] =
-        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-        var result = ""
-        var remainingLength = length
-        
-        while remainingLength > 0 {
-            let randoms: [UInt8] = (0 ..< 16).map { _ in
-                var random: UInt8 = 0
-                let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
-                if errorCode != errSecSuccess {
-                    fatalError(
-                        "Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)"
-                    )
-                }
-                return random
-            }
-            
-            randoms.forEach { random in
-                if remainingLength == 0 {
-                    return
-                }
-                
-                if random < charset.count {
-                    result.append(charset[Int(random)])
-                    remainingLength -= 1
-                }
-            }
-        }
-        
-        return result
-    }
-    
-    @available(iOS 13, *)
-    private func sha256(_ input: String) -> String {
-        let inputData = Data(input.utf8)
-        let hashedData = SHA256.hash(data: inputData)
-        let hashString = hashedData.compactMap {
-            String(format: "%02x", $0)
-        }.joined()
-        
-        return hashString
     }
     
 }
