+++ Project 2 AirNote/SceneDelegate.swift
@@ -16,7 +16,14 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
         // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
-        guard let _ = (scene as? UIWindowScene) else { return }
+        guard let windowScence = (scene as? UIWindowScene) else { return }
+        window = UIWindow(windowScene: windowScence)
+        
+        if FirebaseSignInManager.currentUser == nil {
+            self.window?.rootViewController = UIStoryboard.auth.instantiateInitialViewController()
+        } else {
+            self.window?.rootViewController = UIStoryboard.main.instantiateInitialViewController()
+        }
     }
 
     func sceneDidDisconnect(_ scene: UIScene) {
